slackbot/requirements.txt
requests==2.25.1
SQLAlchemy
psycopg2-binary
pandas